[SCRUM-101] chore: 금액 필드 타입 Int→Long 통일
- OrderGroup·OrderItem·Order·ClaimItem·Seller(baseShippingFee·freeShippingThreshold)·User.pointBalance·PointTransaction.amount를 Long으로 — Product·settlement이 이미 Long이라 정합
- 집계 금액(정산 회차 총액 등)은 Int 상한(21.4억) 초과 가능 → 오버플로우/절단 방지. count·quantity류는 Int 유지
- ERD int 이탈 — DDL 확정(Flyway) 시 BIGINT

+++ domain/src/main/kotlin/com/aechak/domain/order/claim/ClaimItem.kt
@@ -17,7 +17,7 @@ import jakarta.persistence.Table
 class ClaimItem protected constructor(
     orderItem: OrderItem,
     quantity: Int,
-    refundAmount: Int,
+    refundAmount: Long,
 ) : BaseEntity() {
 
     @Id
@@ -32,10 +32,10 @@ class ClaimItem protected constructor(
     val quantity: Int = quantity
 
     @Column(nullable = false)
-    val refundAmount: Int = refundAmount
+    val refundAmount: Long = refundAmount
 
     companion object {
-        fun of(orderItem: OrderItem, quantity: Int, refundAmount: Int): ClaimItem =
+        fun of(orderItem: OrderItem, quantity: Int, refundAmount: Long): ClaimItem =
             ClaimItem(orderItem, quantity, refundAmount)
     }
 }
